fixed unused variables.

+++ cub3d_bonus/src/key_handler_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   key_handler_bonus.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zouddach <zouddach@1337.student.ma>        +#+  +:+       +#+        */
+/*   By: mait-lah <mait-lah@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/09 14:21:27 by zouddach          #+#    #+#             */
-/*   Updated: 2025/02/09 14:52:49 by zouddach         ###   ########.fr       */
+/*   Updated: 2025/02/09 22:54:21 by mait-lah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ bool	check_up_collision(t_game *game, int *new_x, int *new_y)
 	*new_x = game->player.x + game->player.moving_speed * sin(game->player.dir);
 	*new_y = game->player.y + game->player.moving_speed * cos(game->player.dir);
 	if (*new_x < 0 || *new_y < 0 || *new_x >= game->map.rows
-		|| *new_y >= ft_strlen(game->map.map[(int)*new_x]))
+		|| (size_t)(*new_y) >= ft_strlen(game->map.map[(int)*new_x]))
 		return (false);
 	if (game->map.map[(int)*new_x][(int)*new_y] == '1'
 		|| game->map.map[(int)*new_x][(int)*new_y] == 'D')
@@ -45,7 +45,7 @@ bool	check_down_collision(t_game *game, int *new_x, int *new_y)
 	*new_x = game->player.x - game->player.moving_speed * sin(game->player.dir);
 	*new_y = game->player.y - game->player.moving_speed * cos(game->player.dir);
 	if (*new_x < 0 || *new_y < 0 || *new_x >= game->map.rows
-		|| *new_y >= ft_strlen(game->map.map[(int)*new_x]))
+		|| (size_t)(*new_y) >= ft_strlen(game->map.map[(int)*new_x]))
 		return (false);
 	if (game->map.map[(int)*new_x][(int)*new_y] == '1'
 		|| game->map.map[(int)*new_x][(int)*new_y] == 'D')
@@ -73,7 +73,7 @@ bool	check_left_collision(t_game *game, int *new_x, int *new_y)
 	*new_y = game->player.x - game->player.moving_speed * cos(game->player.dir);
 	*new_y = game->player.y + game->player.moving_speed * sin(game->player.dir);
 	if (*new_x < 0 || *new_y < 0 || *new_x >= game->map.rows
-		|| *new_y >= ft_strlen(game->map.map[(int)*new_x]))
+		|| (size_t)(*new_y) >= ft_strlen(game->map.map[(int)*new_x]))
 		return (false);
 	if (game->map.map[(int)*new_y][(int)*new_y] == '1'
 		|| game->map.map[(int)*new_y][(int)*new_y] == 'D')
@@ -99,7 +99,7 @@ bool	check_right_collision(t_game *game, int *new_x, int *new_y)
 	*new_x = game->player.x + game->player.moving_speed * cos(game->player.dir);
 	*new_y = game->player.y - game->player.moving_speed * sin(game->player.dir);
 	if (*new_x < 0 || *new_y < 0 || *new_x >= game->map.rows
-		|| *new_y >= ft_strlen(game->map.map[(int)*new_x]))
+		|| (size_t)(*new_y) >= ft_strlen(game->map.map[(int)*new_x]))
 		return (false);
 	if (game->map.map[(int)*new_x][(int)*new_y] == '1'
 		|| game->map.map[(int)*new_x][(int)*new_y] == 'D')
