+++ cub3d_bonus/src/map_check5_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   map_check5_bonus.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zouddach <zouddach@1337.student.ma>        +#+  +:+       +#+        */
+/*   By: mait-lah <mait-lah@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/09 14:37:39 by zouddach          #+#    #+#             */
-/*   Updated: 2025/02/09 14:58:05 by zouddach         ###   ########.fr       */
+/*   Updated: 2025/02/09 22:47:36 by mait-lah         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ bool	check_wall_collision(t_game *game, int keycode)
 			&new_y))
 		return (false);
 	if (new_x < 0 || new_y < 0 || new_x >= game->map.rows
-		|| new_y >= ft_strlen(game->map.map[(int)new_x]))
+		|| (size_t)new_y >= ft_strlen(game->map.map[(int)new_x]))
 		return (false);
 	if (game->map.map[(int)new_x][(int)new_y] == '1'
 		|| game->map.map[(int)new_x][(int)new_y] == 'D')
